Fix tarball resolution without toplevel `package`
This patch resolves an issue where packages would fail to parse when
they didn't have a toplevel `package` directory inside the tarball.

The fix always removes the toplevel directory, regardless of its name,
which makes the assumption there always is a toplevel directory. Based
on my sample of 100 packages, I have not found any which include the
files directly in the tarball, rather than inside a directory.

Closes #33.