Fix local function reassignment
This changes the approach for finding function accessors by using
recursion when a name wasn't directly defined for the function, to
ensure the access graph goes to the closest possible parent.

It also removes all attempts at evaluating if a function is executed or
not and instead just checks for access.