Fix global variable hoisting
This fixes an issue where assignments without declaration wouldn't be
correctly considered to be part of the global scope, despite JavaScript
hoisting any assignment without declaration by default.

In theory the hoisting only happens when an assignment is executed, so
the following code alone is not sufficient to define the global `test`:

```js
function foo() {
    test = 5;
}
```

However any code executing this function `foo` will permanently hoist
`test` to be a global variable. With our current symbol table
representation, the only viable way to map this is to assume `test` is
always a variable assigned in the global `program` scope.

This hoisting behavior can be suppressed retroactively by declaring the
variable using the `var` keyword. Retroactive declaration does not work
with `let` or `const`. This is why finding the globally hoisted
variables requires finding any potential `var` declarations in the
current scope first.

Closes #42.